organize test facilities with directory
