+++ zingolib/src/test_framework.rs
@@ -1,63 +1,2 @@
 pub(crate) mod macros;
-
-use zcash_primitives::transaction::TxId;
-
-macro_rules! build_method {
-    ($name:ident, $localtype:ty) => {
-        pub fn $name(mut self, $name: $localtype) -> Self {
-            self.$name = Some($name);
-            self
-        }
-    };
-}
-use crate::wallet::notes::TransparentNote;
-pub struct TransparentNoteBuilder {
-    address: Option<String>,
-    txid: Option<TxId>,
-    output_index: Option<u64>,
-    script: Option<Vec<u8>>,
-    value: Option<u64>,
-    spent: Option<Option<(TxId, u32)>>,
-    unconfirmed_spent: Option<Option<(TxId, u32)>>,
-}
-#[allow(dead_code)] //TODO:  fix this gross hack that I tossed in to silence the language-analyzer false positive
-impl TransparentNoteBuilder {
-    pub fn new() -> Self {
-        Self::default()
-    }
-    // Methods to set each field
-    build_method!(address, String);
-    build_method!(txid, TxId);
-    build_method!(output_index, u64);
-    build_method!(script, Vec<u8>);
-    build_method!(value, u64);
-    build_method!(spent, Option<(TxId, u32)>);
-    build_method!(unconfirmed_spent, Option<(TxId, u32)>);
-
-    // Build method
-    pub fn build(self) -> TransparentNote {
-        TransparentNote::from_parts(
-            self.address.unwrap(),
-            self.txid.unwrap(),
-            self.output_index.unwrap(),
-            self.script.unwrap(),
-            self.value.unwrap(),
-            self.spent.unwrap(),
-            self.unconfirmed_spent.unwrap(),
-        )
-    }
-}
-
-impl Default for TransparentNoteBuilder {
-    fn default() -> Self {
-        TransparentNoteBuilder {
-            address: Some("default_address".to_string()),
-            txid: Some(TxId::from_bytes([0u8; 32])),
-            output_index: Some(0),
-            script: Some(vec![]),
-            value: Some(0),
-            spent: Some(None),
-            unconfirmed_spent: Some(None),
-        }
-    }
-}
+pub(crate) mod mocks;
