+++ zingolib/src/wallet/data.rs
@@ -704,7 +704,7 @@ pub use crate::wallet::transaction_record::TransactionRecord;
 fn single_transparent_note_makes_is_incoming_true() {
     // A single transparent note makes is_incoming_trsaction true.
     let txid = TxId::from_bytes([0u8; 32]);
-    let transparent_note = crate::test_framework::TransparentNoteBuilder::new()
+    let transparent_note = crate::test_framework::mocks::TransparentNoteBuilder::new()
         .address("t".to_string())
         .spent(Some((txid, 3)))
         .build();
